pyproject.toml
[tool.poetry]
name = "wind"
version = "0.1.0"
description = ""
authors = ["Foppe Hemminga <foppe@fop.pe>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
python-dotenv = "^0.21.1"
psycopg = "^3.1.8"
pytz = "^2022.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
